backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
httpx==0.27.0
websockets==12.0
colorlog==6.8.2
python-dotenv==1.0.1
networkx==3.3
